slightly closer to writing asteroid support
